refactor: trazendo login e logout para o controller.
